Add models for World, Room, Channel

See merge request virtual-events/venueless!10

+++ video/server/stayseated/core/services/chat.py
@@ -1,9 +1,10 @@
+from contextlib import suppress
+
 from channels.db import database_sync_to_async
 from django.db import IntegrityError, transaction
 from django.db.models import Max
 
-from ..models import ChatEvent, User
-from ..models.chat import Membership
+from ..models import Channel, ChatEvent, Membership, User
 from ..serializers.chat import ChatEventSerializer
 from ..utils.redis import aioredis
 from .user import get_public_users, get_user_by_id
@@ -15,18 +16,20 @@ class ChatService:
 
     @database_sync_to_async
     def get_channels_for_user(self, user_id, is_volatile=None):
-        qs = Membership.objects.filter(world_id=self.world_id, user_id=user_id,)
+        qs = Membership.objects.filter(
+            channel__world_id=self.world_id, user_id=user_id,
+        )
         if is_volatile is not None:  # pragma: no cover
             qs = qs.filter(volatile=is_volatile)
-        return list(qs.values_list("channel", flat=True))
+        return list(str(channel) for channel in qs.values_list("channel", flat=True))
 
     async def get_channel_users(self, channel):
         users = await get_public_users(
             # We're doing an ORM query in an async method, but it's okay, since it is not going to be evaluated but
             # lazily passed to get_public_users which will use it as a subquery :)
-            ids=Membership.objects.filter(
-                world_id=self.world_id, channel=channel
-            ).values_list("user_id", flat=True),
+            ids=Membership.objects.filter(channel_id=channel).values_list(
+                "user_id", flat=True
+            ),
             world_id=self.world_id,
         )
         return users
@@ -43,20 +46,19 @@ class ChatService:
     @database_sync_to_async
     def membership_is_volatile(self, channel, uid):
         try:
-            m = Membership.objects.get(
-                world_id=self.world_id, channel=channel, user_id=uid,
-            )
+            m = Membership.objects.get(channel=channel, user_id=uid)
             return m.volatile
         except Membership.DoesNotExist:  # pragma: no cover
             return False
 
     @database_sync_to_async
-    def add_channel_user(self, channel, uid, volatile):
-        try:
+    def add_channel_user(self, channel_id, uid, volatile):
+        # Currently, users are undeletable, so this should be a pretty impossible code path. Anyway, if it happens,
+        # there is probably no harm in ignoring it.
+        with suppress(User.DoesNotExist):
             with transaction.atomic():
                 m, created = Membership.objects.get_or_create(
-                    world_id=self.world_id,
-                    channel=channel,
+                    channel=Channel.objects.get(id=channel_id),
                     user=User.objects.get(id=uid, world_id=self.world_id),
                     defaults={"volatile": volatile},
                 )
@@ -64,30 +66,23 @@ class ChatService:
                     m.volatile = False
                     m.save(update_fields=["volatile"])
                 return created
-        except User.DoesNotExist:  # pragma: no cover
-            # Currently, users are undeletable, so this should be a pretty impossible code path. Anyway, if it happens,
-            # there is probably no harm in ignoring it.
-            return False
 
     @database_sync_to_async
-    def remove_channel_user(self, channel, uid):
-        Membership.objects.filter(
-            world_id=self.world_id, channel=channel, user_id=uid,
-        ).delete()
+    def remove_channel_user(self, channel_id, uid):
+        Membership.objects.filter(channel_id=channel_id, user_id=uid,).delete()
 
     @database_sync_to_async
     def get_events(self, channel, before_id, count=50):
-        events = ChatEvent.objects.filter(
-            id__lt=before_id, world_id=self.world_id, channel=channel,
-        ).order_by("-id")[: min(count, 1000)]
+        events = ChatEvent.objects.filter(id__lt=before_id, channel=channel,).order_by(
+            "-id"
+        )[: min(count, 1000)]
         return ChatEventSerializer(reversed(list(events)), many=True).data
 
     @database_sync_to_async
-    def _store_event(self, channel, id, event_type, content, sender):
+    def _store_event(self, channel_id, id, event_type, content, sender):
         ce = ChatEvent.objects.create(
             id=id,
-            world_id=self.world_id,
-            channel=channel,
+            channel=Channel.objects.get(id=channel_id),
             event_type=event_type,
             content=content,
             sender=sender,
@@ -110,7 +105,7 @@ class ChatService:
             event_id = await redis.incr("chat.event_id")
         try:
             return await self._store_event(
-                channel=channel,
+                channel_id=channel,
                 id=event_id,
                 event_type=event_type,
                 content=content,
